requirements.txt
streamlit
pandas
langchain
langchain-community
langchain-huggingface
langchain-google-genai
faiss-cpu
sentence-transformers
python-dotenv
